backend/requirements.txt
asgiref==3.11.1
certifi==2025.11.12
charset-normalizer==3.4.3
Django==6.0.3
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.3.0
idna==3.10
packaging==26.0
PyJWT==2.12.1
requests==2.32.5
sqlparse==0.5.5
urllib3==2.5.0
uv==0.8.22
